implement repository interface for company and contact person
