+++ app/Crm/Contacts/ContactInterface.php
@@ -11,8 +11,8 @@ interface ContactInterface
      * @return mixed
      */
     public function index();
-    public function store(StoreContactRequest $request);
+    public function store(Request $request);
     public function edit($id);
     public function update(StoreContactRequest $request,$id);
-    public function destroy(Request $request);
+    public function destroy($request);
 }
