+++ app/Http/Controllers/Api/ApiResponseTrait.php
@@ -4,6 +4,12 @@ namespace App\Http\Controllers\Api;
 
 trait ApiResponseTrait
 {
+    /**
+     * @param $data
+     * @param $messsage
+     * @param $status
+     * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\Routing\ResponseFactory|\Illuminate\Http\Response
+     */
     public function apiResponse($data = null , $messsage = null , $status = null)
     {
         $array = [
