+++ app/Http/Controllers/CompanyController.php
@@ -3,20 +3,27 @@
 namespace App\Http\Controllers;
 
 use App\Http\Requests\StoreCompanyRequest;
+use Crm\Companies\CompanyInterface;
 use Crm\Companies\Models\Company;
+use Crm\Contacts\ContactInterface;
 use Illuminate\Http\Request;
 
 class CompanyController extends Controller
 {
     /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
+     * @param CompanyInterface $company
+     */
+    public function __construct(CompanyInterface $company)
+    {
+        $this->companyInterface = $company;
+    }
+
+    /**
+     * @return mixed
      */
     public function index()
     {
-        $company = Company::all();
-        return view('layouts.companies.index',compact('company'));
+        return $this->companyInterface->index();
     }
 
     /**
@@ -30,37 +37,17 @@ class CompanyController extends Controller
     }
 
     /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
+     * @param Request $request
+     * @return mixed
      */
     public function store(Request $request)
     {
-        /*
-         * Move Logo
-         */
-        $logo = $request -> file('logo');
-        $ext = $logo->getClientOriginalExtension();
-        $name = "companies-".uniqid() . ".$ext";
-        $logo -> move( public_path('storage/') , $name);
-
-            Company::create([
-                'name' => $request->name,
-                'email' => $request->email,
-                'logo' => $name,
-                'url' => $request->url
-            ]);
-            session()->flash('Add', __('Companies Data Added Successfully') );
-            return redirect()->back();
-
+        return $this->companyInterface->store($request);
     }
 
     /**
-     * Display the specified resource.
-     *
-     * @param  \Crm\Companies\Models\Company  $company
-     * @return \Illuminate\Http\Response
+     * @param Company $company
+     * @return void
      */
     public function show(Company $company)
     {
@@ -68,63 +55,30 @@ class CompanyController extends Controller
     }
 
     /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  \Crm\Companies\Models\Company  $company
-     * @return \Illuminate\Http\Response
+     * @param $id
+     * @return mixed
      */
     public function edit($id)
     {
-        $company = Company::findorfail($id);
-        return view ('layouts.companies.edit',compact('company'));
+        return $this->companyInterface->index($id);
     }
 
     /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \Crm\Companies\Models\Company  $company
-     * @return \Illuminate\Http\Response
+     * @param StoreCompanyRequest $request
+     * @param $id
+     * @return mixed
      */
     public function update(StoreCompanyRequest $request, $id)
     {
-        $company = Company::findorFail($request->id);
-
-        $name = $company -> logo;
-
-        if($request->hasFile('logo'))
-        {
-
-            $logo = $request -> file('logo');
-            $ext = $logo->getClientOriginalExtension();
-            $name = "companies-".uniqid() . ".$ext";
-            $logo -> move( public_path('storage') , $name);
-        }
-            $company->update([
-                'name' => $request->name,
-                'email' => $request->email,
-                'logo' => $name ,
-                'url' => $request->url
-            ]);
-            session()->flash('Edit','Updated Successfully');
-            return redirect()->back();
+        return $this->companyInterface->update($request,$id);
     }
 
     /**
-     * Remove the specified resource from storage.
-     *
-     * @param  \Crm\Companies\Models\Company  $company
-     * @return \Illuminate\Http\Response
+     * @param $companyId
+     * @return mixed
      */
-    public function destroy(Request $request)
+    public function destroy($companyId)
     {
-        try {
-            Company::destroy($request->company_id);
-            session()->flash('Deleted', 'Data has been deleted successfully From Companies');
-            return redirect()->back();
-        }
-        catch (\Exception $e) {
-            return redirect()->back()->withErrors(['error' => $e->getMessage()]);
-        }
+       return $this->companyInterface->destroy($companyId);
     }
 }
